broke out the context store into its own file
